version using 16-bit primes
I had to comment out an assert(dropped.len() == num_drop), I think it
probably had a collision.

+++ quack-bm/src/construct.rs
@@ -8,7 +8,7 @@ use multiset::HashMultiSet;
 use sha2::{Digest, Sha256};
 
 fn benchmark_construct_strawman1(
-    numbers: Vec<u64>,
+    numbers: Vec<u16>,
     num_packets: usize,
     num_drop: usize,
 ) -> Duration {
@@ -40,7 +40,7 @@ fn benchmark_construct_strawman1(
 }
 
 fn benchmark_construct_strawman2(
-    numbers: Vec<u64>,
+    numbers: Vec<u16>,
     num_packets: usize,
     num_drop: usize,
 ) -> Duration {
@@ -61,7 +61,7 @@ fn benchmark_construct_strawman2(
 }
 
 fn benchmark_construct_power_sum_32(
-    numbers: Vec<u64>,
+    numbers: Vec<u16>,
     size: usize,
     num_packets: usize,
     num_drop: usize,
@@ -89,7 +89,7 @@ fn benchmark_construct_power_sum_32(
     let t2 = Instant::now();
 
     let duration = t2 - t1;
-    info!("Insert {} numbers into 2 Quacks (u64, \
+    info!("Insert {} numbers into 2 Quacks (u16, \
         threshold = {}): {:?}", num_packets, size, duration);
     duration
 }
@@ -112,7 +112,7 @@ pub fn run_benchmark(
     let mut durations: Vec<Duration> = vec![];
 
     for i in 0..(num_trials + 1) {
-        let numbers: Vec<u64> =
+        let numbers: Vec<u16> =
             (0..(num_packets + 10)).map(|_| rng.gen()).collect();
 
         let duration = match quack_ty {
